// src/app/models/types/game-level.ts
// export interface GameLevel {}

export enum GameLevel {
  LETTER = 'lettering',
  WORD = 'word',
  PARAGRAPH = 'paragraph',
  STORY = 'story',
  NUMBER_RECOGNITION_ONE = 'number-recognition-one',
  NUMBER_RECOGNITION_TWO = 'number-recognition-two',
  NUMBER_RECOGNITION_THREE = 'number-recognition-three',
  PLACE_VALUE = 'place-value',
}